sonar fixes
