+++ firstvoices/backend/tests/test_apis/test_dictionary_cleanup_api.py
@@ -9,6 +9,7 @@ from backend.tests import factories
 from backend.tests.test_apis.base_api_test import (
     BaseReadOnlyUncontrolledSiteContentApiTest,
     SiteContentCreateApiTestMixin,
+    SuperAdminAsyncJobPermissionsMixin,
     WriteApiTestMixin,
 )
 from backend.tests.test_search_indexing.base_indexing_tests import (
@@ -20,6 +21,7 @@ class TestDictionaryCleanupAPI(
     WriteApiTestMixin,
     SiteContentCreateApiTestMixin,
     TransactionOnCommitMixin,
+    SuperAdminAsyncJobPermissionsMixin,
     BaseReadOnlyUncontrolledSiteContentApiTest,
 ):
     API_LIST_VIEW = "api:dictionary-cleanup-list"
@@ -196,148 +198,6 @@ class TestDictionaryCleanupAPI(
                 },
             ],
         }
-
-    @pytest.mark.django_db
-    def test_list_403_non_member(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        self.client.force_authenticate(user=user)
-
-        response = self.client.get(self.get_list_endpoint(site_slug=site.slug))
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    @pytest.mark.parametrize(
-        "role",
-        [Role.ASSISTANT, Role.EDITOR, Role.LANGUAGE_ADMIN],
-    )
-    def test_list_empty_non_superuser(self, role):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        factories.MembershipFactory.create(user=user, site=site, role=role)
-        self.create_minimal_instance(site=site, visibility=None)
-        self.client.force_authenticate(user=user)
-
-        response = self.client.get(self.get_list_endpoint(site_slug=site.slug))
-        response_data = json.loads(response.content)
-
-        assert response.status_code == 200
-        assert response_data["count"] == 0
-        assert response_data["results"] == []
-
-    @pytest.mark.django_db
-    def test_list_empty_staff(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        factories.AppMembershipFactory.create(user=user, role=AppRole.STAFF)
-        self.client.force_authenticate(user=user)
-
-        response = self.client.get(self.get_list_endpoint(site_slug=site.slug))
-        response_data = json.loads(response.content)
-
-        assert response.status_code == 200
-        assert response_data["count"] == 0
-        assert response_data["results"] == []
-
-    @pytest.mark.django_db
-    def test_get_403_non_member(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        self.client.force_authenticate(user=user)
-
-        instance = self.create_minimal_instance(site=site, visibility=Visibility.PUBLIC)
-
-        response = self.client.get(
-            self.get_detail_endpoint(
-                key=self.get_lookup_key(instance), site_slug=site.slug
-            )
-        )
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    @pytest.mark.parametrize(
-        "role",
-        [Role.MEMBER, Role.ASSISTANT, Role.EDITOR, Role.LANGUAGE_ADMIN],
-    )
-    def test_get_403_non_superuser(self, role):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        self.client.force_authenticate(user=user)
-        instance = self.create_minimal_instance(site=site, visibility=None)
-
-        response = self.client.get(
-            self.get_detail_endpoint(
-                key=self.get_lookup_key(instance), site_slug=site.slug
-            )
-        )
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    def test_get_403_staff(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        factories.AppMembershipFactory.create(user=user, role=AppRole.STAFF)
-        self.client.force_authenticate(user=user)
-        instance = self.create_minimal_instance(site=site, visibility=None)
-
-        response = self.client.get(
-            self.get_detail_endpoint(
-                key=self.get_lookup_key(instance), site_slug=site.slug
-            )
-        )
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    def test_post_403_non_member(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        self.client.force_authenticate(user=user)
-
-        data = {}
-
-        response = self.client.post(
-            self.get_list_endpoint(site_slug=site.slug), data=data, format="json"
-        )
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    @pytest.mark.parametrize(
-        "role",
-        [Role.MEMBER, Role.ASSISTANT, Role.EDITOR, Role.LANGUAGE_ADMIN],
-    )
-    def test_post_403_non_superuser(self, role):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        factories.MembershipFactory.create(user=user, site=site, role=role)
-        self.client.force_authenticate(user=user)
-
-        data = {}
-
-        response = self.client.post(
-            self.get_list_endpoint(site_slug=site.slug), data=data, format="json"
-        )
-
-        assert response.status_code == 403
-
-    @pytest.mark.django_db
-    def test_post_403_staff(self):
-        site = factories.SiteFactory.create()
-        user = factories.UserFactory.create()
-        factories.AppMembershipFactory.create(user=user, role=AppRole.STAFF)
-        self.client.force_authenticate(user=user)
-
-        data = {}
-
-        response = self.client.post(
-            self.get_list_endpoint(site_slug=site.slug), data=data, format="json"
-        )
-
-        assert response.status_code == 403
 
     @pytest.mark.django_db
     def test_clear(self, is_preview=False):
